pyproject.toml
[tool.isort]
profile = "black"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''
[tool.poetry]
name = "scrapper-fii-py"
version = "0.1.0"
description = "Python FII scrapper"
authors = ["Khwesten Heiner <k-heiner@hotmail.com>"]
readme = "README.md"
packages = [{include = "scrapper"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["standard"], version = "^0.115.4"}
pydantic = "^2.9.2"
starlette = "^0.41.2"
aiodns = "^3.2.0"
aiofiles = "^24.1.0"
aiohttp = "^3.10.10"
lxml = "^5.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
autoflake = "^2.3.1"
isort = "^5.13.2"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
factory-boy = "^3.3.1"
faker = "^30.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
